+++ js/constants.js
@@ -1,0 +1,37 @@
+const QUANTITY_PHOTOS = 25;//относится к id,url
+const MIN_LIKES = 15;
+const MAX_LIKES = 200;
+const MIN_COMMENTS = 0;
+const MAX_COMMENTS = 30;
+const QUANTITY_AVATAR = 6;
+
+const NAMES = [
+  'Илья',
+  'Варвара',
+  'Иван',
+  'Александр',
+  'Петр',
+  'Юлия',
+  'Алексей',
+  'Софья',
+  'Михаил',
+  'Леонид',
+  'Мария',
+  'Семен',
+  'Елена',
+  'Екатерина',
+  'Николай',
+];
+
+const DESCRIPTION = ['Луна','Февраль','Поле','Розовый закат','Кот','Каменистые горы'];
+
+const MESSAGES = [
+  'Всё отлично!',
+  'В целом всё неплохо. Но не всё.',
+  'Когда вы делаете фотографию, хорошо бы убирать палец из кадра. В конце концов это просто непрофессионально.',
+  'Моя бабушка случайно чихнула с фотоаппаратом в руках и у неё получилась фотография лучше.',
+  'Я поскользнулся на банановой кожуре и уронил фотоаппарат на кота и у меня получилась фотография лучше.',
+  'Лица у людей на фотке перекошены, как будто их избивают. Как можно было поймать такой неудачный момент?!',
+];
+
+export {QUANTITY_PHOTOS, MIN_LIKES, MAX_LIKES, MIN_COMMENTS, MAX_COMMENTS, QUANTITY_AVATAR, NAMES, DESCRIPTION, MESSAGES};
