+++ js/functions.js
@@ -1,3 +1,35 @@
+
+// Делу — время
+// Напишите функцию, которая принимает время начала и конца рабочего дня, а также время старта и продолжительность встречи в минутах и возвращает true,
+
+//если встреча не выходит за рамки рабочего дня, и false, если выходит.
+
+// Время указывается в виде строки в формате часы:минуты. Для указания часов и минут могут использоваться как две цифры, так и одна. Например, 8 часов 5 минут могут быть указаны по-разному: 08:05, 8:5, 08:5 или 8:05.
+
+// Продолжительность задаётся числом. Гарантируется, что и рабочий день, и встреча укладываются в одни календарные сутки.
+/*
+'8:00' - начало рабочего дня
+'17:30' - конец рабочего дня
+'14:00' - начало встречи
+90 - продолжительность встречи в минутах
+*/
+// имяФункции('08:00', '17:30', '14:00', 90); // true
+// имяФункции('8:0', '10:0', '8:0', 120);     // true
+// имяФункции('08:00', '14:30', '14:00', 90); // false
+// имяФункции('14:00', '17:30', '08:0', 90);  // false
+// имяФункции('8:00', '17:30', '08:00', 900); // false
+
+function getParametrOfMeeting (startOfWorkingDay, endOfWorkingDay, meetingStartTime, durationOfMeeting) {
+  const arrFromData = [[...startOfWorkingDay.split(':')], [...endOfWorkingDay.split(':')], [...meetingStartTime.split(':')]];
+  const newArr = arrFromData.map((item) => Number(item[0]) * 60 + Number(item[1]));
+  [startOfWorkingDay, endOfWorkingDay, meetingStartTime] = newArr;
+  return (startOfWorkingDay <= meetingStartTime) && (meetingStartTime + durationOfMeeting <= endOfWorkingDay);
+}
+
+getParametrOfMeeting ('14:00', '17:30', '08:0', 90);
+
+//getParametrOfMeeting ('14:00', '17:30', '08:0', 90);
+
 // // // // const checkLength = (string, maxLength) => string?.length <= maxLength;
 // // // // checkLength();
 
@@ -117,3 +149,4 @@
 // }
 
 // topSalaries (salaries);
+
