+++ js/utils.js
@@ -1,0 +1,33 @@
+function getRandomInteger (min,max) {
+  const lower = Math.ceil(Math.min(Math.abs(min),Math.abs(max)));
+  const upper = Math.floor(Math.max(Math.abs(min),Math.abs(max)));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+}
+//Количество фотографий
+const generateCountOfPhotos = () => {
+  let count = 0;
+  return function () {
+    return ++count;
+  };
+};
+
+//создать уникальный id
+const createUniqueId = (min,max) => {
+  const previousValues = [];
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min,max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+};
+
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, generateCountOfPhotos, getRandomArrayElement, createUniqueId};
